pyproject.toml
[project]
name = "secrets-manager"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-resource-manager>=1.14.2",
    "google-cloud-secret-manager>=2.23.3",
    "pydantic>=2.11.4",
    "textual>=3.1.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.11.7",
]

[tool.pytest]
testpaths = ["tests"]
python_files = ["test_*.py"]
pythonpath = ["."]

[tool.pytest.ini_options]
addopts = "-v --tb=short"
testpaths = ["tests"]
